pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "git-pulse"
version = "0.1.0"
description = "Analyze git repo history for development hotspots with LLM-powered insights"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
dependencies = [
    "typer>=0.9.0",
    "gitpython>=3.1.0",
    "litellm>=1.0.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
git-pulse = "git_pulse.cli:app"

[tool.pytest.ini_options]
testpaths = ["tests"]
